Move Article creation styles to external file
